requirements.txt
kafka-python==2.0.2
apache-airflow==2.7.0
cassandra-driver==3.28.0
pyspark==3.4.1
